% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_IW.R, R/scale_color_IW.R,
%   R/scale_fill_IW.R, R/theme__plot_IW.R
\name{ggsave_IW_plot}
\alias{ggsave_IW_plot}
\alias{scale_color_IW}
\alias{scale_fill_IW}
\alias{theme_IW}
\title{Ein Paket für R-Grafiken im Farbschema des Instituts der deutschen Wirtschaft}
\usage{
ggsave_IW_plot(plot_object = last_plot(), filename)

scale_color_IW(...)

scale_fill_IW(...)

theme_IW(base_size = 14)
}
\arguments{
\item{hist_data}{A table of weather variables with PWS created by hist_data function}

\item{variable}{A character string of variable name}
}
\description{
Ein Paket für R-Grafiken im Farbschema des Instituts der deutschen Wirtschaft

Ein Paket für R-Grafiken im Farbschema des Instituts der deutschen Wirtschaft

Ein Paket für R-Grafiken im Farbschema des Instituts der deutschen Wirtschaft

Ein Paket für R-Grafiken im Farbschema des Instituts der deutschen Wirtschaft
}
\examples{
check
check
check
check
}
